one unique constant file
