+++ python-lib/dku_tools.py
@@ -1,7 +1,7 @@
 import dataiku
 from dataiku.customrecipe import get_input_names_for_role, get_output_names_for_role
 
-from design_experiment.constants import SizeDefinition, AttributionMethod, Parameters
+from constants import SizeDefinition, AttributionMethod, Parameters
 
 
 def get_input_output() -> tuple:
