--- python-lib/global_constants.py
@@ -1,10 +1,0 @@
-from enum import Enum
-
-
-class Columns(object):
-    AB_GROUP = "dku_ab_group"
-
-
-class Group(Enum):
-    A = "A"
-    B = "B"
