--- python-lib/results/constants.py
@@ -1,6 +1,0 @@
-from global_constants import Columns
-
-
-class StatisticsColumns(Columns):
-    SUCCESS_RATE = "success_rate"
-    SIZE = "sample_size"
